local font

+++ src/main/java/com/kobe/warehouse/service/facturation/dto/TiersPayantDossierFactureDto.java
@@ -1,4 +1,3 @@
 package com.kobe.warehouse.service.facturation.dto;
 
-public record TiersPayantDossierFactureDto(Long id, String name, Integer totalAmount, int factureItemCount) {
-}
+public record TiersPayantDossierFactureDto(Long id, String name, Integer totalAmount, Long factureItemCount) {}
